more documentation and file format
