+++ Documentation/App Mockup/tipoPrediccion.txt
@@ -1,0 +1,513 @@
+/* TipoPrediccion */
+
+position: relative;
+width: 1440px;
+height: 1024px;
+
+background: #F4F9FF;
+box-shadow: 4px 4px 4px rgba(68, 28, 107, 0.25), 0px 4px 4px rgba(68, 28, 107, 0.5), 0px 4px 10px rgba(68, 28, 107, 0.3);
+
+
+/* Rectangle 3 */
+
+position: absolute;
+width: 1440px;
+height: 150px;
+left: 0px;
+top: 0px;
+
+background: #8F88E0;
+border: 2px solid rgba(0, 0, 0, 0.5);
+
+
+/* Rectangle 2 */
+
+position: absolute;
+width: 350px;
+height: 1024px;
+left: 0px;
+top: 0px;
+
+background: #CBCBEA;
+
+
+/* Rectangle 4 */
+
+position: absolute;
+width: 1030px;
+height: 800px;
+right: 30px;
+bottom: 44px;
+
+background: #FAFCFF;
+border: 5px solid rgba(141, 99, 174, 0.2);
+box-sizing: border-box;
+
+
+/* btnSig */
+
+position: absolute;
+width: 330px;
+height: 130px;
+left: 730px;
+top: 720px;
+
+
+
+/* Rectangle 1 */
+
+position: absolute;
+left: 0%;
+right: 0%;
+top: 0%;
+bottom: 0%;
+
+background: rgba(144, 135, 239, 0.9);
+border: 2px solid rgba(116, 96, 242, 0.2);
+box-sizing: border-box;
+box-shadow: inset 5px 8px 4px rgba(255, 255, 255, 0.2);
+border-radius: 20px;
+
+
+/* Siguiente */
+
+position: absolute;
+left: 3.93%;
+right: 3.63%;
+top: 13.39%;
+bottom: 14.17%;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 36px;
+line-height: 42px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #21202B;
+
+
+
+/* btnSig */
+
+position: absolute;
+width: 330px;
+height: 130px;
+left: 730px;
+top: 540px;
+
+
+
+/* Rectangle 1 */
+
+position: absolute;
+left: 0%;
+right: 0%;
+top: 0%;
+bottom: 0%;
+
+background: rgba(144, 135, 239, 0.9);
+border: 2px solid rgba(116, 96, 242, 0.2);
+box-sizing: border-box;
+box-shadow: inset 5px 8px 4px rgba(255, 255, 255, 0.2);
+border-radius: 20px;
+
+
+/* Siguiente */
+
+position: absolute;
+left: 3.93%;
+right: 3.63%;
+top: 13.39%;
+bottom: 14.17%;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 36px;
+line-height: 42px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #21202B;
+
+
+
+/* btnSig */
+
+position: absolute;
+width: 330px;
+height: 130px;
+left: 730px;
+top: 370px;
+
+
+
+/* Rectangle 1 */
+
+position: absolute;
+left: 0%;
+right: 0%;
+top: 0%;
+bottom: 0%;
+
+background: rgba(144, 135, 239, 0.9);
+border: 2px solid rgba(116, 96, 242, 0.2);
+box-sizing: border-box;
+box-shadow: inset 5px 8px 4px rgba(255, 255, 255, 0.2);
+border-radius: 20px;
+
+
+/* Siguiente */
+
+position: absolute;
+left: 3.93%;
+right: 3.63%;
+top: 13.39%;
+bottom: 14.17%;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 36px;
+line-height: 42px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #21202B;
+
+
+
+/* rbtnDosData */
+
+position: absolute;
+width: 400px;
+height: 80px;
+left: 695px;
+top: 260px;
+
+
+
+/* Selecione el tipo de predicción del conjunto de datos suministrado */
+
+position: absolute;
+width: 400px;
+height: 80px;
+left: 695px;
+top: 260px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 24px;
+line-height: 28px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #696969;
+
+
+
+/* btnInfo2 */
+
+position: absolute;
+width: 50px;
+height: 50px;
+left: 1130px;
+top: 410px;
+
+
+
+/* Ellipse 2 */
+
+position: absolute;
+width: 50px;
+height: 50px;
+left: 1130px;
+top: 410px;
+
+background: rgba(101, 85, 198, 0.2);
+border: 2px solid #6555C6;
+box-sizing: border-box;
+
+
+/* i */
+
+position: absolute;
+width: 25px;
+height: 44.55px;
+left: 1142.5px;
+top: 412.73px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 48px;
+line-height: 56px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #6555C6;
+
+
+
+/* btnInfo1 */
+
+position: absolute;
+width: 50px;
+height: 50px;
+left: 1130px;
+top: 579px;
+
+
+
+/* Ellipse 2 */
+
+position: absolute;
+width: 50px;
+height: 50px;
+left: 1130px;
+top: 579px;
+
+background: rgba(101, 85, 198, 0.2);
+border: 2px solid #6555C6;
+box-sizing: border-box;
+
+
+/* i */
+
+position: absolute;
+width: 25px;
+height: 44.55px;
+left: 1142.5px;
+top: 581.73px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 48px;
+line-height: 56px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #6555C6;
+
+
+
+/* btnInfo1 */
+
+position: absolute;
+width: 50px;
+height: 50px;
+left: 1130px;
+top: 759px;
+
+
+
+/* Ellipse 2 */
+
+position: absolute;
+width: 50px;
+height: 50px;
+left: 1130px;
+top: 759px;
+
+background: rgba(101, 85, 198, 0.2);
+border: 2px solid #6555C6;
+box-sizing: border-box;
+
+
+/* i */
+
+position: absolute;
+width: 25px;
+height: 44.55px;
+left: 1142.5px;
+top: 761.73px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 48px;
+line-height: 56px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #6555C6;
+
+
+
+/* [LOGO] */
+
+position: absolute;
+visibility: hidden;
+width: 196px;
+height: 115px;
+left: 77px;
+top: 35px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 24px;
+line-height: 28px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #000000;
+
+
+
+/* PASO 4 - Tipo de predicción */
+
+position: absolute;
+width: 638px;
+height: 71px;
+left: 437px;
+top: 41px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: bold;
+font-size: 36px;
+line-height: 42px;
+display: flex;
+align-items: center;
+
+color: #FFFFFF;
+
+
+
+/* bxs-left-arrow-circle 1 */
+
+position: absolute;
+width: 70px;
+height: 70px;
+left: 425px;
+top: 225px;
+
+
+
+/* Vector */
+
+position: absolute;
+left: 8.33%;
+right: 8.33%;
+top: 8.33%;
+bottom: 8.33%;
+
+background: #9087EF;
+
+
+/* VoorspellingLogo 2 */
+
+position: absolute;
+width: 230px;
+height: 199px;
+left: 60px;
+top: 25px;
+
+background: url(VoorspellingLogo.png);
+
+
+/* Group 2 */
+
+position: absolute;
+width: 311px;
+height: 717px;
+left: 17px;
+top: 278px;
+
+
+
+/* Iván-David Rey-Rueda buc16372023@mail.udes.edu.co German Diaz-Figueredo buc17372054@mail.udes.edu.co */
+
+position: absolute;
+width: 306px;
+height: 126px;
+left: 22px;
+top: 497px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 18px;
+line-height: 21px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #696969;
+
+
+
+/* Voorspelling, aplicación de escritorio para aprendizaje de máquina supervisado */
+
+position: absolute;
+width: 306px;
+height: 126px;
+left: 17px;
+top: 278px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 18px;
+line-height: 21px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #696969;
+
+
+
+/* Universidad de Santander Ingeniería de Software 2020 */
+
+position: absolute;
+width: 306px;
+height: 126px;
+left: 22px;
+top: 727px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 18px;
+line-height: 21px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #696969;
+
+
+
+/* Version 1.0 Licencia BSD 3 */
+
+position: absolute;
+width: 306px;
+height: 126px;
+left: 22px;
+top: 869px;
+
+font-family: Roboto;
+font-style: normal;
+font-weight: normal;
+font-size: 14px;
+line-height: 16px;
+display: flex;
+align-items: center;
+text-align: center;
+
+color: #696969;
+
